# docs/swagger.yaml
basePath: /
definitions:
...
  enums.OrderStatus:
    enum:
    - pending_payment
    - paid
    - processing
    - shipped
    - delivered
    - cancelled
    type: string
    x-enum-varnames:
    - StatusPendingPayment
    - StatusPaid
    - StatusProcessing
    - StatusShipped
    - StatusDelivered
    - StatusCancelled
...
  order.CreateOrderRequest:
    properties:
      delivery_date:
        type: string
      items:
        items:
          $ref: '#/definitions/order.OrderItemInput'
        type: array
      pickup_point:
        example: Mirzo Ulug'bek. Buyuk Ipak Yoli st. 109/45
        type: string
    required:
    - items
    - pickup_point
    type: object
  order.Order:
    properties:
      created_at:
        type: string
      delivery_date:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/order.OrderItem'
        type: array
      order_date:
        type: string
      pickup_point:
        type: string
      receipt_url:
        type: string
      status:
        $ref: '#/definitions/enums.OrderStatus'
      total_amount:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  order.OrderItem:
    properties:
      price:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
      total_price:
        type: integer
    type: object
  order.OrderItemInput:
    properties:
      price:
        example: 12000
        type: integer
      product_id:
        type: integer
      quantity:
        example: 4
        type: integer
    required:
    - price
    - product_id
    - quantity
    type: object
  order.UpdateStatusRequest:
    properties:
      status:
        example: processing
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: REST API для управления заказами, продуктами и пользователями.
  license:
    name: MIT
  termsOfService: http://example.com/terms/
  title: Order Service API
  version: "1.0"
paths:
...
  /api/v1/orders:
    get:
      description: Возвращает список заказов текущего пользователя
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/order.Order'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all user orders
      tags:
      - orders
  /api/v1/orders/:
    post:
      consumes:
      - application/json
      description: Create a new order for authenticated user
      parameters:
      - description: Order info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/order.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: order_id
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: insufficient stock
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create order
      tags:
      - Orders
  /api/v1/orders/{id}:
    delete:
      description: Удаляет заказ по ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete order by ID (admin)
      tags:
      - orders
    get:
      description: Возвращает заказ по ID (user/admin)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.Order'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - orders
    put:
      description: Обновляет статус заказа по ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: New status
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/order.UpdateStatusRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update order status (admin)
      tags:
      - orders
  /api/v1/orders/{id}/cancel:
    get:
      description: Отмена заказа пользователем
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel order
      tags:
      - orders
  /api/v1/orders/export:
    get:
      description: Экспорт заказов в JSON по фильтрам
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by min amount
        in: query
        name: min_amount
        type: integer
      - description: Filter by max amount
        in: query
        name: max_amount
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Export orders (admin)
      tags:
      - orders
  /api/v1/orders/export/csv:
    get:
      description: Экспорт заказов в CSV по фильтрам
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by min amount
        in: query
        name: min_amount
        type: integer
      - description: Filter by max amount
        in: query
        name: max_amount
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: CSV file
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Export orders as CSV (admin)
      tags:
      - orders
  /api/v1/orders/stats:
    get:
      description: Получает статистику заказов по статусам за период
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get order stats (admin)
      tags:
      - orders
...
swagger: "2.0"